Add files via upload
Apache Spark is one of the most used Big Data tools in recent years. In this, I am uploading the Pyspark files which I have worked on SQL functions and pandas Datasets.
Using the Python framework in Apache Spark makes tougher problems simple and optimized with the help of numerous in-built functions.